requirements.txt
bidict==0.21.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dataclasses==0.8
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SocketIO==5.2.0
GDAL @ file:///.\GDAL-3.0.4-cp36-cp36m-win_amd64.whl
h11==0.13.0
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
LEPL==5.1.3
MarkupSafe==2.0.1
numpy==1.19.5
python-engineio==4.3.2
python-socketio==5.6.0
requests==2.27.1
rfc6266==0.0.4
scipy==1.5.4
simple-websocket==0.5.2
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wsproto==1.0.0
zipp==3.6.0
jsonschema>=4.0.0
pytest>=6.0.0
pyyaml>=5.0.0
